Major overhaul, should have branched.
Luckily I managed to finish this branch before urgent changes were
required

Basically, the views and interactive menu is now loaded with ajax.

This allows the settings to persist from one view to the next.

There is a nice tabbed menu on the left that allows the user to set
filters, and preferences. These settings persist through page-loads.